Refactor and improve how config and options are read
