+++ package.json
@@ -80,7 +80,6 @@
     "chalk": "^2.4.2",
     "find-up": "^3.0.0",
     "inquirer": "^6.2.2",
-    "joi": "^14.3.1",
     "lodash.flatten": "^4.4.0",
     "lodash.get": "^4.4.2",
     "lodash.isempty": "^4.4.0",
@@ -96,7 +95,6 @@
     "@types/find-up": "^2.1.1",
     "@types/inquirer": "^6.0.0",
     "@types/jest": "^24.0.11",
-    "@types/joi": "^14.3.2",
     "@types/lodash.flatten": "^4.4.6",
     "@types/lodash.get": "^4.4.6",
     "@types/lodash.isempty": "^4.4.6",
