--- src/lib/errors.ts
@@ -1,7 +1,0 @@
-export class HandledError extends Error {
-  constructor(message: string) {
-    super(message);
-    Error.captureStackTrace(this, HandledError);
-    this.name = 'HandledError';
-  }
-}
