+++ src/lib/git.ts
@@ -2,7 +2,7 @@ import childProcess from 'child_process';
 import rimraf from 'rimraf';
 import * as env from './env';
 import { mkdirp, stat, exec } from './rpc';
-import { HandledError } from './errors';
+import { HandledError } from './HandledError';
 
 async function folderExists(path: string): Promise<boolean> {
   try {
