+++ src/types/commons.d.ts
@@ -1,0 +1,5 @@
+export type PromiseReturnType<Func> = Func extends (
+  ...args: any[]
+) => Promise<infer Value>
+  ? Value
+  : Func;
